requirements.txt
Flask==3.0.3
gunicorn==21.2.0
google-cloud-storage==2.17.0
google-cloud-bigquery==3.25.0
requests==2.32.3
